Play timed text even after 35.79 minutes
int overflow when converting current position in ms to Us

+++ media/libmediaplayerservice/nuplayer/NuPlayer.cpp
@@ -2234,7 +2234,7 @@ void NuPlayer::onSourceNotify(const sp<AMessage> &msg) {
             int posMs;
             int64_t timeUs, posUs;
             driver->getCurrentPosition(&posMs);
-            posUs = posMs * 1000;
+            posUs = (int64_t) posMs * 1000ll;
             CHECK(buffer->meta()->findInt64("timeUs", &timeUs));
 
             if (posUs < timeUs) {
